requirements.txt
discord.py==2.2.2
selenium-wire==5.1.0
python-dotenv==1.0.0